Track entities and components, and fix all memory leaks.
- all entities destroyed at program end.
- fontstash fixed to not memory leak when using the freetype library.
- components should now specialise CleanupComponent instead of RemoveComponent, if they need to clean things up. This will be run when component changes are committed, between frames, rather than when the remove request is made.

+++ src/components/Game/UIComponents.h
@@ -10,8 +10,8 @@ namespace Game::UI
 	struct LoadingScreen
 	{
 		std::string m_nextLoadedFilename{ "assets/preload.res" };
-		int m_currentlyLoaded{ 0 };
-		int m_totalToLoad{ 0 };
+		usize m_currentlyLoaded{ 0 };
+		usize m_totalToLoad{ 0 };
 		Core::Resource::SpriteID m_fullScreenSprite;
 	};
